requirements.txt
matplotlib
seaborn
plotly
bokeh
altair
plotnine
pygal
holoviews
vispy
mpld3
pandas
numpy
IPython
lxml
PyQt6
altair_saver
selenium 
pillow
vl-convert-python
